Authorization

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,23 +1,26 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
     public interface IRestaurantService
     {
-        int Create(CreateRestaurantDto dto);
+        int Create(CreateRestaurantDto dto/*, int userId*/);
         IEnumerable<RestaurantDto> GetAll();
         RestaurantDto GetById(int id);
-        bool Delete(int id);
+        bool Delete(int id/*, ClaimsPrincipal user*/);
         
-        void /*bool*/ Update(int id, UpdateRestaurantDto dto);
+        void /*bool*/ Update(int id, UpdateRestaurantDto dto/*, ClaimsPrincipal user*/);
     }
 
     public class RestaurantService : IRestaurantService
@@ -26,12 +29,17 @@ namespace RestaurantAPI.Services
         private readonly IMapper _mapper;
 
         public readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService, IUserContextService userContextService)
         {
+            
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public RestaurantDto GetById(int id)
         {
@@ -51,14 +59,15 @@ namespace RestaurantAPI.Services
             return restaurantsDto;
         }
 
-        public int Create(CreateRestaurantDto dto)
+        public int Create(CreateRestaurantDto dto/*, int userId*/)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
         }
-        public bool Delete(int id)
+        public bool Delete(int id/*, ClaimsPrincipal user*/)
         {
             _logger.LogInformation("loguje delete");
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
@@ -66,17 +75,28 @@ namespace RestaurantAPI.Services
             {
                 return false;
             }
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
             return true;
         }
-        public void /*bool*/ Update(int id, UpdateRestaurantDto dto)
+        public void /*bool*/ Update(int id, UpdateRestaurantDto dto/*, ClaimsPrincipal user*/)
         {
+            
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
             if (restaurant is null)
             {
                 throw new NotFoundException("Restaurant not found");
                 //return false;
+            }
+            var authorizationResult =  _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
             }
             restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
